renamed ops agent plugin dir
