+++ builds/ops_agent_plugin.sh
@@ -17,4 +17,4 @@ set -x -e
 DESTDIR=$1
 mkdir -p "$DESTDIR/opt/google-cloud-ops-agent/libexec"
 go build -buildvcs=false -o "$DESTDIR/opt/google-cloud-ops-agent/libexec/google_cloud_ops_agent_uap_plugin" \
-  github.com/GoogleCloudPlatform/ops-agent/cmd/ops_agent_uap_wrapper
+  github.com/GoogleCloudPlatform/ops-agent/cmd/ops_agent_uap_plugin
